feat: 턴시스템 승리조건/패배조건 완료

+++ Assets/1_Scripts/2_World/City.cs
@@ -43,7 +43,7 @@ public class City : MonoBehaviour
 
     public float PL {
         get {
-            return PA / GameInfo.maxPA;
+            return PA / (float)GameInfo.maxPA;
         }
     }
